+++ lib/dockerfile_opam.mli
@@ -28,10 +28,6 @@ module RPM : sig
 
   val install_system_opam : [< `CentOS6 | `CentOS7 ] -> Dockerfile.t
   (** Install the system OPAM packages via Yum *)
-
-  val add_opensuse_repo : [< `CentOS6 | `CentOS7 ] -> Dockerfile.t
-  (** Add a Yum entry for the OpenSUSE Build Service OCaml repo *)
-
 end
 
 (** Apt distribution specific rules involving OPAM and OCaml *)
@@ -46,13 +42,7 @@ module Apt : sig
 
     val install_system_opam : Dockerfile.t
     (** Install the system OPAM packages via [apt-get] *)
-
-    val add_opensuse_repo :
-      [< `Debian of [< `Stable | `Testing ]
-       | `Ubuntu of [< `V14_04 | `V14_10 ] ] -> Dockerfile.t
-    (** Add a [apt] repository entry for the OpenSUSE Build
-        Service repository for OCaml *)
-  end
+end
 
 val run_as_opam : ('a, unit, string, Dockerfile.t) format4 -> 'a
 (** [run_as_opam fmt] runs the command specified by the [fmt]
@@ -73,9 +63,6 @@ val install_opam_from_source : ?prefix:string -> ?branch:string -> unit -> Docke
 (** Commands to install OPAM via a source code checkout from GitHub.
     The [branch] defaults to the [1.2] stable branch.
     The binaries are installed under [<prefix>/bin], defaulting to [/usr/local/bin]. *)
-
-val install_ext_plugin : Dockerfile.t
-(** Installs the external dependency plugin shell script. *)
 
 val header: string -> string -> Dockerfile.t
 (** [header image tag] initalises a fresh Dockerfile using the [image:tag]