+++ lib/dockerfile.ml
@@ -167,7 +167,7 @@ module Linux = struct
   (** Debian rules *)
   module Apt = struct
     let update = run "apt-get -y update" @@ run "apt-get -y upgrade"
-    let install fmt = ksprintf (run "apt-get -y install %s") fmt
+    let install fmt = ksprintf (run "DEBIAN_FRONTEND=noninterative apt-get -y install %s") fmt
 
     let dev_packages ?extra () =
       update @@