+++ CHANGES.md
@@ -4,6 +4,7 @@
   to use Docker Hub dynamic branch support to build all permutations.
 * Correctly escape the `run_exec`, `entrypoint_exec` and `cmd_exec`
   JSON arrays so that the strings are quoted.
+* Run `yum clean` after a Yum installation.
 
 1.1.1 (2015-03-11):
 * Add a `?prefix` argument to `install_opam_from_source`