+++ CHANGES.md
@@ -6,6 +6,7 @@
   JSON arrays so that the strings are quoted.
 * Run `yum clean` after a Yum installation.
 * Add support for Alpine Linux.
+* Cleanup OPAM build directory to save container space after building from source.
 
 1.1.1 (2015-03-11):
 * Add a `?prefix` argument to `install_opam_from_source`