+++ CHANGES.md
@@ -1,4 +1,5 @@
 1.2.0:
+* Remove `onbuild` triggers from OPAM containers as it inhibits caching (suggestion via @talex5)
 * Include specific Debian versions (v7,8,9) in addition to the stable/unstable streams.
 * Add `Dockerfile.crunch` to reduce the number of layers by combining
   repeated `RUN` commands.